// AppFaktum/src/app/models/sucursal-cliente.model.ts
export interface ISucursalCliente{
    suclDepto: string;
    suclCiudad: string;
    suclCodigo: string;
    suclContacto: string;
    suclCorreo: string;
    suclDiasPago: number;
    suclListaPrecio: string;
    suclNombre: string;
    suclTelefono: string;
    suclClienteId: number;
    id: number;
    estado: number;
    fechaCreacion: string;
    fechaModificacion: string;
}